+++ cli/src/lib.rs
@@ -1,6 +1,7 @@
 pub mod api_client;
 pub mod print;
-pub mod runner;
+pub mod quarantine;
 pub mod scanner;
+pub mod test;
 pub mod upload;
 pub mod validate;
